feat: add the second stage
Now I looked at the code for the webserver.go and realized how the
application is supposed to work.

Actually, there is more a separation of concerns.  The webserver
just deals with, well... webserving.

The web assets are configured to the webserver in the second stage,
and they can be easily changed if needed, with no rebuild if we have
to change some plain file or image that will be served.